requirements.txt
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
openpyxl>=3.1.0
xlrd>=2.0.0
python-dateutil>=2.8.0
pytz>=2023.3
requests>=2.31.0
urllib3>=2.0.0
cryptography>=41.0.0
python-dotenv>=1.0.0 